Toggle to start nomad
The toggle `consul_started` (default to `true`) can help when
creating images. If set to false during image build time the
handlers will not be triggered, and the service will not be
started (although enabled at boot).

This prevents the population of the data dir with information
that might not be accurate when creating an instance based on
an image